+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Added support for `XPENDING` command (#1558)
 - Added support for `XSETID` command (#1559)
 - Added validation and support for the new `BITOP` command operations (#1566)
+- Added test validation for the new SVS-VAMANA vector index type for FT.CREATE (#1578)
 
 ### Changed
 - Handle and retry `LOADING` errors from Sentinel replicas (#1536)